pyproject.toml
[project]
name = "road-gen-ai"
version = "0.1.0"
description = "AI-powered OpenDRIVE road generation tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "scenariogeneration>=0.9.0",
    "pydantic>=2.0.0",
    "litellm==1.80.5",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.57.0",
    "qdrant-client>=1.16.2",
    "scikit-learn>=1.7.2",
]

[project.scripts]
road-gen-ai = "road_gen_ai.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
